fix clear state, remove unused files
