--- lua/hyper/components/request.lua
@@ -1,1 +1,0 @@
--- displays the request buffer
